#116: Refactoring into smaller components
